requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
lxml
werkzeug
reportlab
python-barcode
Pillow
python-dotenv
